UC2_AddContact

+++ AddressBook/Contact.cs
@@ -15,7 +15,7 @@ namespace AddressbookSystem
         string city;
         string state;
         int zip;
-        int phoneNumber;
+        long phoneNumber;
         string email;
 
         /// <summary>
@@ -29,7 +29,7 @@ namespace AddressbookSystem
         /// <param name="zip">The zip.</param>
         /// <param name="phoneNumber">The phone number.</param>
         /// <param name="email">The email.</param>
-        public Contact(string firstName, string lastName, string address, string city, string state, int zip, int phoneNumber, string email)
+        public Contact(string firstName, string lastName, string address, string city, string state, int zip, long phoneNumber, string email)
         {
             this.firstName = firstName;
             this.lastName = lastName;
